medium_crawler/requirements.txt
Django==4.2.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
python-dateutil==2.8.2
django-bootstrap4==23.2
django-crispy-forms==2.1
crispy-bootstrap4==2022.1